feat(brain-core): extract brain primitives into khive-brain-core crate
Foundation for dissolving pack-brain: pure types (BetaPosterior,
SectionType, SectionPosteriorState, ProfileRecord, BrainState,
FeedbackSignal/EventKind) moved to a new foundation-level crate
with no runtime/storage dependencies. Pack-brain and pack-memory
now import from khive-brain-core directly.

Closes the type-sharing prerequisite from issue #5. Next: dissolve
pack-brain's feedback/fold into consuming packs.
