+++ crates/Cargo.toml
@@ -19,6 +19,7 @@ members = [
     "khive-pack-kg",
     "khive-pack-gtd",
     "khive-pack-memory",
+    "khive-brain-core",
     "khive-pack-brain",
     "khive-pack-comm",
     "khive-pack-schedule",
