+++ crates/khive-brain-core/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "khive-brain-core"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "Brain primitives — Beta posteriors, section types, profile state, weight derivation"
+
+[dependencies]
+serde = { workspace = true }
+serde_json = { workspace = true }
+uuid = { workspace = true }
+chrono = { workspace = true }
+rand = { workspace = true }
