+++ crates/khive-pack-brain/src/lib.rs
@@ -4,8 +4,6 @@ pub mod event;
 pub mod fold;
 pub mod handlers;
 pub mod persist;
-pub mod section;
-pub mod state;
 pub mod tunable;
 
 mod pack;
